Frankly, it doesn't make sense for vm_pageout_wakeup_thresh to have a negative
value (it is only ever set to a fraction of v_free_min, which is unsigned and
also obviously non-negative).  But I'm not going to try and convert every
non-negative scalar in the VM to unsigned today, so just cast it for the
comparison.

Submitted by:	Clang 3.3
Sponsored by:	EMC / Isilon Storage Division

+++ sys/sys/vmmeter.h
@@ -186,7 +186,7 @@ int
 vm_paging_needed(void)
 {
     return (vm_cnt.v_free_count + vm_cnt.v_cache_count <
-        vm_pageout_wakeup_thresh);
+        (u_int)vm_pageout_wakeup_thresh);
 }
 
 #endif
